# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "OM PAY API",
        "description": "API de paiement mobile OM PAY avec authentification OTP et gestion de comptes",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Serveur de développement"
        },
        {
            "url": "https://ompay-isuf.onrender.com/api",
            "description": "Serveur de production"
        }
    ],
    "paths": {
...
        "/comptes/{numero_compte}/transactions/paiement": {
            "post": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Effectuer un paiement marchand",
                "description": "Effectue un paiement vers un marchand en utilisant son code marchand depuis le compte spécifié",
                "operationId": "paiement",
                "parameters": [
                    {
                        "name": "numero_compte",
                        "in": "path",
                        "description": "Numéro du compte émetteur",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "OM-2025-AB12-CD34"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "montant"
                                ],
                                "properties": {
                                    "code_marchand": {
                                        "description": "Code unique du marchand (optionnel si telephone_marchand est fourni)",
                                        "type": "string",
                                        "example": "MARCHAND001"
                                    },
                                    "telephone_marchand": {
                                        "description": "Numéro de téléphone du marchand (optionnel si code_marchand est fourni)",
                                        "type": "string",
                                        "example": "781562042"
                                    },
                                    "montant": {
                                        "description": "Montant du paiement",
                                        "type": "number",
                                        "format": "float",
                                        "example": 2500
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Paiement effectué avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Paiement effectué avec succès"
                                        },
                                        "data": {
                                            "properties": {
                                                "transaction": {
                                                    "properties": {
                                                        "reference": {
                                                            "type": "string",
                                                            "example": "TXN-20251110-ABC123"
                                                        },
                                                        "type": {
                                                            "type": "string",
                                                            "example": "paiement"
                                                        },
                                                        "montant": {
                                                            "type": "number",
                                                            "format": "float",
                                                            "example": 2500
                                                        },
                                                        "frais": {
                                                            "type": "number",
                                                            "format": "float",
                                                            "example": 0
                                                        },
                                                        "statut": {
                                                            "type": "string",
                                                            "example": "reussi"
                                                        },
                                                        "compteEmetteur": {
                                                            "properties": {
                                                                "numero_compte": {
                                                                    "type": "string",
                                                                    "example": "OM-2025-AB12-CD34"
                                                                }
                                                            },
                                                            "type": "object"
                                                        },
                                                        "marchand": {
                                                            "properties": {
                                                                "raison_sociale": {
                                                                    "type": "string",
                                                                    "example": "Boutique Express"
                                                                },
                                                                "code_marchand": {
                                                                    "type": "string",
                                                                    "example": "MARCHAND001"
                                                                }
                                                            },
                                                            "type": "object"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erreur de traitement (solde insuffisant, montant invalide, etc.)"
                    },
                    "403": {
                        "description": "Accès non autorisé"
                    },
                    "404": {
                        "description": "Marchand ou compte non trouvé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/comptes/{numero_compte}/transactions": {
            "get": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Liste des transactions du compte",
                "description": "Retourne la liste de toutes les transactions du compte spécifié",
                "operationId": "getTransactions",
                "parameters": [
                    {
                        "name": "numero_compte",
                        "in": "path",
                        "description": "Numéro du compte",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "OM-2025-AB12-CD34"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des transactions",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "transactions": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "reference": {
                                                                "type": "string",
                                                                "example": "TXN-20251110-ABC123"
                                                            },
                                                            "type": {
                                                                "type": "string",
                                                                "enum": [
                                                                    "paiement",
                                                                    "transfert",
                                                                    "depot"
                                                                ],
                                                                "example": "transfert"
                                                            },
                                                            "montant": {
                                                                "type": "number",
                                                                "format": "float",
                                                                "example": 5000
                                                            },
                                                            "frais": {
                                                                "type": "number",
                                                                "format": "float",
                                                                "example": 50
                                                            },
                                                            "statut": {
                                                                "type": "string",
                                                                "enum": [
                                                                    "reussi",
                                                                    "echec"
                                                                ],
                                                                "example": "reussi"
                                                            },
                                                            "created_at": {
                                                                "type": "string",
                                                                "format": "date-time",
                                                                "example": "2025-11-10T09:00:00Z"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Accès non autorisé"
                    },
                    "404": {
                        "description": "Aucun compte trouvé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}